# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AERANERD API",
        "description": "API para gerenciamento de áreas, matérias e escritores do sistema AERANERD",
        "contact": {
            "email": "contato@aeranerd.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/areas": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Listar todas as áreas",
                "description": "Retorna uma lista paginada de todas as áreas disponíveis ordenadas alfabeticamente",
                "operationId": "f16518e18a005a2f966b63c0e1745342",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de áreas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Area"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Could not retrieve areas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Areas"
                ],
                "summary": "Criar uma nova área",
                "description": "Cria uma nova área no sistema",
                "operationId": "beeb06a4f56c4e31c8448c405163bed9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vchr_AreaNome"
                                ],
                                "properties": {
                                    "vchr_AreaNome": {
                                        "description": "Nome da área",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Quadrinhos"
                                    },
                                    "vchr_Tag": {
                                        "description": "Tag da área",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "comics"
                                    },
                                    "type": {
                                        "description": "Tipo da área",
                                        "type": "string",
                                        "enum": [
                                            "bd",
                                            "pasta"
                                        ],
                                        "example": "bd"
                                    },
                                    "b_menu": {
                                        "description": "Se aparece no menu",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "int_rolePermission": {
                                        "description": "Permissão de role",
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Área criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Area"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Could not create area."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/areas/{id}": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Obter uma área específica",
                "description": "Retorna os detalhes de uma área específica pelo ID",
                "operationId": "d45f3822a5bdba2f44520ac32857bfb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Área encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Area"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Área não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Area not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Could not retrieve area."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Areas"
                ],
                "summary": "Atualizar uma área existente",
                "description": "Atualiza os dados de uma área específica",
                "operationId": "cc20a1af054acc1ea386968c8d9ba095",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vchr_AreaNome"
                                ],
                                "properties": {
                                    "vchr_AreaNome": {
                                        "description": "Nome da área",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Quadrinhos"
                                    },
                                    "vchr_Tag": {
                                        "description": "Tag da área",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "comics"
                                    },
                                    "type": {
                                        "description": "Tipo da área",
                                        "type": "string",
                                        "enum": [
                                            "bd",
                                            "pasta"
                                        ],
                                        "example": "bd"
                                    },
                                    "b_menu": {
                                        "description": "Se aparece no menu",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "int_rolePermission": {
                                        "description": "Permissão de role",
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Área atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Area"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Área não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Area not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Could not update area."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Areas"
                ],
                "summary": "Excluir uma área",
                "description": "Remove uma área do sistema",
                "operationId": "820be3b502aa09c30f43b877c1c4527e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Área excluída com sucesso"
                    },
                    "404": {
                        "description": "Área não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Area not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Could not delete area."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Area": {
                "title": "Area",
                "description": "Modelo de Área",
                "properties": {
                    "int_Id": {
                        "description": "ID da área",
                        "type": "integer"
                    },
                    "vchr_AreaNome": {
                        "description": "Nome da área",
                        "type": "string"
                    },
                    "vchr_Tag": {
                        "description": "Tag da área",
                        "type": "string"
                    },
                    "type": {
                        "description": "Tipos da área",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "b_menu": {
                        "description": "Se aparece no menu",
                        "type": "boolean"
                    },
                    "int_rolePermission": {
                        "description": "Permissão de role",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}